fix: fix size & duration components support
fix: fix size components missing

fix: fix duration support enum value

refactor: reuse image Select component

fix: fix default model
