+++ src/routes/(main)/(create)/video/features/PromptInput/index.tsx
@@ -21,6 +21,7 @@ import {
   InlineVideoFrames,
 } from '@/routes/(main)/(create)/features/GenerationInput';
 import { AspectRatioSelect } from '@/routes/(main)/(create)/image/features/ConfigPanel';
+import Select from '@/routes/(main)/(create)/image/features/ConfigPanel/components/Select';
 import VideoModelItem from '@/routes/(main)/(create)/video/features/ConfigPanel/components/ModelSelect/VideoModelItem';
 import { aiProviderSelectors, useAiInfraStore } from '@/store/aiInfra';
 import { useUserStore } from '@/store/user';
@@ -51,6 +52,23 @@ const AspectRatioItem = memo(() => {
   return <AspectRatioSelect options={options} value={value} onChange={(v) => setValue(v as any)} />;
 });
 
+const SizeItem = memo(() => {
+  const { value, setValue, enumValues } = useVideoGenerationConfigParam('size');
+
+  const options = useMemo(
+    () =>
+      enumValues?.map((size) => ({
+        label: size,
+        value: size,
+      })) ?? [],
+    [enumValues],
+  );
+
+  if (options.length === 0) return null;
+
+  return <Select options={options} value={value} onChange={setValue} />;
+});
+
 const ResolutionItem = memo(() => {
   const { value, setValue, enumValues } = useVideoGenerationConfigParam('resolution');
   const options = useMemo(
@@ -73,7 +91,31 @@ const ResolutionItem = memo(() => {
 });
 
 const DurationItem = memo(() => {
-  const { value, setValue, min, max, step } = useVideoGenerationConfigParam('duration');
+  const { value, setValue, min, max, step, enumValues } = useVideoGenerationConfigParam('duration');
+
+  const options = useMemo(
+    () =>
+      enumValues && enumValues.length > 0
+        ? enumValues.map((v) => ({
+            label: String(v),
+            value: v,
+          }))
+        : [],
+    [enumValues],
+  );
+
+  if (options.length > 0) {
+    return (
+      <Segmented
+        block
+        options={options}
+        style={{ width: '100%' }}
+        value={value ?? min}
+        variant="filled"
+        onChange={(v) => setValue(Number(v) as any)}
+      />
+    );
+  }
 
   return (
     <SliderWithInput
@@ -143,6 +185,7 @@ const PromptInput = ({ showTitle = false }: PromptInputProps) => {
   const isSupportEndImageUrl = useVideoStore(isSupportedParamSelector('endImageUrl'));
   const isSupportAspectRatio = useVideoStore(isSupportedParamSelector('aspectRatio'));
   const isSupportResolution = useVideoStore(isSupportedParamSelector('resolution'));
+  const isSupportSize = useVideoStore(isSupportedParamSelector('size'));
   const isSupportDuration = useVideoStore(isSupportedParamSelector('duration'));
   const isSupportSeed = useVideoStore(isSupportedParamSelector('seed'));
   const isSupportGenerateAudio = useVideoStore(isSupportedParamSelector('generateAudio'));
@@ -267,6 +310,12 @@ const PromptInput = ({ showTitle = false }: PromptInputProps) => {
                       <Flexbox gap={6}>
                         <Text fontSize={12}>{t('config.resolution.label')}</Text>
                         <ResolutionItem />
+                      </Flexbox>
+                    )}
+                    {isSupportSize && (
+                      <Flexbox gap={6}>
+                        <Text fontSize={12}>{t('config.size.label')}</Text>
+                        <SizeItem />
                       </Flexbox>
                     )}
                     {isSupportSeed && (
